rasterizer_cache: Move texture filtering to the runtime

+++ src/video_core/renderer_opengl/texture_filters/anime4k/anime4k_ultrafast.cpp
@@ -39,7 +39,7 @@
 
 namespace OpenGL {
 
-Anime4kUltrafast::Anime4kUltrafast(u16 scale_factor) : TextureFilterBase(scale_factor) {
+Anime4kUltrafast::Anime4kUltrafast(u32 scale_factor) : TextureFilterBase(scale_factor) {
     const OpenGLState cur_state = OpenGLState::GetCurState();
 
     vao.Create();
