+++ src/video_core/renderer_opengl/texture_filters/bicubic/bicubic.h
@@ -14,7 +14,7 @@ class Bicubic : public TextureFilterBase {
 public:
     static constexpr std::string_view NAME = "Bicubic";
 
-    explicit Bicubic(u16 scale_factor);
+    explicit Bicubic(u32 scale_factor);
     void Filter(const OGLTexture& src_tex, Common::Rectangle<u32> src_rect,
                 const OGLTexture& dst_tex, Common::Rectangle<u32> dst_rect) override;
